+++ test/index.ts
@@ -3,7 +3,8 @@ import {readFileSync} from 'fs';
 import 'mocha';
 import {resolve} from 'path';
 import postcss from 'postcss';
-import stylusParser from '../src/index';
+import stylus, {RendererOptions} from 'stylus';
+import getParser from '../src/index';
 
 describe(
 	'Parse Stylus',
@@ -13,17 +14,30 @@ describe(
 			'should parse stylus file',
 			() => check( 'simple' ),
 		);
+		it(
+			'should resolve url',
+			() => check(
+				'resolve-url',
+				{
+					'resolve url': true,
+					'functions': {
+						url: stylus.resolver( {nocheck: true} ),
+					},
+				},
+			),
+		);
 	},
 );
 
 /**
  * Check PostCSS result with specific options.
  * 
- * @param options Plugin options.
- * @param input Input CSS.
- * @param output Expected output CSS.
+ * @param options Stylus render options.
  */
-function check( name: string ): Promise<void>
+function check(
+	name: string,
+	options: Partial<RendererOptions> = {},
+): Promise<void>
 {
 	const sourceName = resolve( __dirname, `fixtures/${name}.styl` );
 	const expectedName = resolve( __dirname, `fixtures/${name}.css` );
@@ -33,7 +47,7 @@ function check( name: string ): Promise<void>
 			readFileSync( sourceName, 'utf8' ),
 			{
 				from: sourceName,
-				parser: stylusParser,
+				parser: getParser( options ),
 			},
 		)
 		.then(