+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+- Ability to specify options for Stylus render function.
+
+### Changed
+
+- Parser function now returned as result of execution of the package function.
+
 ## [1.0.0] - 2018-10-14
 
 ### Changed